requirements.txt
aiohttp==3.6.3
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
confusable-homoglyphs==3.2.0
Django==3.1.1
django-appconf==1.0.4
django-crispy-forms==1.9.2
django-phonenumber-field==5.0.0
django-registration==3.1.1
idna==2.10
maxminddb==2.0.3
multidict==4.7.6
phonenumbers==8.12.9
psycopg2==2.8.6
python-dotenv==0.15.0
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.11
win32-setctime==1.0.2
yadisk==1.2.14
yarl==1.5.1
